requirements.txt
requests
lxml
python-dotenv
psycopg2-binary
